requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
paho-mqtt==2.1.0
pika==1.3.2
pluggy==1.5.0
pymongo==4.8.0
pytest==8.3.2
Werkzeug==3.0.3
